Disable the FIDO U2F provider if the WebAuthn 2FA plugin is activated.
The WebAuthn 2FA plugin basically supercedes the functionality of the
FIDO U2F provider, but doesn't disable the FIDO U2F provider yet.

The 2FA plugin also plans to remove the FIDO U2F provider in a future
release, so let's get ahead of them now. This might get reverted if
the FIDO U2F provider is disabled when that happens.