Revalidation adjustments for two-factor v0.11.0.
We need to manually display the revalidation block above the 2FA table
due to how we are hiding the entire 2FA table when revalidation is
required. (See related commit 27eac38.)

We are also removing the 'Configure a primary two-factor method along
with a backup method, such as Recovery Codes...' string because a
notice is already displayed when a backup method is missing.